made the last final changes before adding the readme file to submit it. I cleaned up my functions file and now its easier to read. every file is included on the functions file
