+++ wp-content/plugins/simple_property_listing/js/myjs.js
@@ -69,11 +69,6 @@ $(function() {
 });
 
 
-$(".chosen").chosen({
-  no_results_text: "Oops, nothing found!",
-  max_selected_options: 2,
-});
-
 
 });
